v1.0.4 - fix for pages
Extension should now work on pages and profiles.

Note: Because of the timing of events the page must be fully loaded
before you attempt to share something. If Post Editor doesn't display
then just click cancel and try to share again and it will display if the
page has completely loaded.